Create solution.sql
First, get the city and population divided by area as the density. Then, use the rank window function of population/area (since density does not exist in the subquery yet) by descending order, as well as by ascending order from the cities population table, where the area is not 0 (to avoid divide by 0 errors). From this table, select the city and density where either the descending or ascending ranks are equal to 1 (which gives the maximum and minimum values.