Create solution.sql
get the subquery with the table with distinct created_at dates greater than 1 and distinct product IDs greater than 1, with the user ID/product ID combination not being in the rank 1 created date (this excludes the first date). Now, from this subquery table, selected the count of distinct user IDs.