Create solution.sql
First, create the subquery with distinct ID, user ID, created at date over the partition of user ID and type, with the max value of the date used to identify the most recent date. The condition is to get the type being 'Refinance'. Then, this table is joined with the submissions table on the loan ID, where the most recent date is the created at date.